FWSECURITY-1003: Added Release build configuration, reworked CI config, added BlinkyApp to non-secure CY8CPROTO-062-4343W board

See merge request repo/cy_mcuboot_project/cy_mcuboot!59

+++ boot/cypress/CypressBootloader/source/main.c
@@ -245,7 +245,7 @@ int main(void)
     // FWSECURITY-676
     if(0 != rc)
     {
-        BOOT_LOG_ERR("Policy parsing failed with code 0x%08x", rc);
+        BOOT_LOG_ERR("Policy parsing failed with code 0x%08x", (int)rc);
     }
     else /*    if(0 == rc) */
     {
